Rename namespace elasticsearch6 -> elasticsearch
