+++ elasticsearch/compat.py
