+++ elasticsearch/connection/base.py
