+++ elasticsearch/connection/http_urllib3.py
