+++ elasticsearch/connection/pooling.py
