+++ elasticsearch/connection_pool.py
