+++ elasticsearch/exceptions.py
