+++ elasticsearch/helpers/errors.py
