+++ elasticsearch/helpers/test.py
@@ -7,15 +7,15 @@ try:
 except ImportError:
     from unittest import TestCase, SkipTest
 
-from elasticsearch6 import Elasticsearch
-from elasticsearch6.exceptions import ConnectionError
+from elasticsearch import Elasticsearch
+from elasticsearch.exceptions import ConnectionError
 
 
 def get_test_client(nowait=False, **kwargs):
     # construct kwargs from the environment
     kw = {"timeout": 5}
     if "TEST_ES_CONNECTION" in os.environ:
-        from elasticsearch6 import connection
+        from elasticsearch import connection
 
         kw["connection_class"] = getattr(connection, os.environ["TEST_ES_CONNECTION"])
