+++ example/queries.py
@@ -4,7 +4,7 @@ from __future__ import print_function
 import logging
 from dateutil.parser import parse as parse_date
 
-from elasticsearch6 import Elasticsearch
+from elasticsearch import Elasticsearch
 
 
 def print_search_stats(results):
