+++ test_elasticsearch/test_connection_pool.py
@@ -1,11 +1,11 @@
 import time
 
-from elasticsearch6.connection_pool import (
+from elasticsearch.connection_pool import (
     ConnectionPool,
     RoundRobinSelector,
     DummyConnectionPool,
 )
-from elasticsearch6.exceptions import ImproperlyConfigured
+from elasticsearch.exceptions import ImproperlyConfigured
 
 from .test_cases import TestCase
