+++ test_elasticsearch/test_exceptions.py
@@ -1,4 +1,4 @@
-from elasticsearch6.exceptions import TransportError
+from elasticsearch.exceptions import TransportError
 
 from .test_cases import TestCase
