+++ test_elasticsearch/test_server/__init__.py
@@ -1,4 +1,4 @@
-from elasticsearch6.helpers.test import (
+from elasticsearch.helpers.test import (
     get_test_client,
     ElasticsearchTestCase as BaseTestCase,
 )
