+++ test_elasticsearch/test_server/test_common.py
@@ -9,9 +9,9 @@ from os.path import exists, join, dirname, pardir
 import yaml
 from shutil import rmtree
 
-from elasticsearch6 import TransportError
-from elasticsearch6.compat import string_types
-from elasticsearch6.helpers.test import _get_version
+from elasticsearch import TransportError
+from elasticsearch.compat import string_types
+from elasticsearch.helpers.test import _get_version
 
 from ..test_cases import SkipTest
 from . import ElasticsearchTestCase
