+++ test_elasticsearch/test_server/test_helpers.py
@@ -1,4 +1,4 @@
-from elasticsearch6 import helpers, TransportError
+from elasticsearch import helpers, TransportError
 
 from . import ElasticsearchTestCase
 from ..test_cases import SkipTest
