+++ tox.ini
@@ -15,15 +15,15 @@ deps =
     black
 commands =
     black --target-version=py27 \
-        elasticsearch6/ \
+        elasticsearch/ \
         test_elasticsearch/ \
         setup.py
     black --target-version=py27 --check \
-        elasticsearch6/ \
+        elasticsearch/ \
         test_elasticsearch/ \
         setup.py
     flake8 \
-        elasticsearch6/ \
+        elasticsearch/ \
         test_elasticsearch/
 
 [testenv:docs]
